+++ shared/src/utils/stats/ContributionStatsCalculator.ts
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { DateTime } from 'luxon';
 import { FirestoreAdmin } from '../../firebase/FirestoreAdmin';
 import { Contribution, CONTRIBUTION_FIRESTORE_PATH, StatusKey, User, USER_FIRESTORE_PATH } from '../../types';
-import { getLatestExchangeRate } from '../currency';
+import { getLatestExchangeRate } from '../exchangeRates';
 import { cumulativeSum, StatsEntry } from './utils';
 import Timestamp = firestore.Timestamp;
