+++ shared/src/utils/stats/PaymentStatsCalculator.ts
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { DateTime } from 'luxon';
 import { FirestoreAdmin } from '../../firebase/FirestoreAdmin';
 import { Payment, PAYMENT_FIRESTORE_PATH } from '../../types';
-import { getLatestExchangeRate } from '../currency';
+import { getLatestExchangeRate } from '../exchangeRates';
 import { cumulativeSum, groupByAndSort, StatsEntry } from './utils';
 import Timestamp = firestore.Timestamp;
